Add READY and WAITING states in position mode
We want to avoid issues with users setting large integral coefficients,
which cause position bounds to be violated if the disk is dithering on
entry to the position mode. With the new waiting state, after trying a
step, a user can put the device into wait mode, which turns off PID
control, zeros errors, and awaits the next command. It goes to WAITING
while the disk stops, then to READY once disk is zeroed.